Completes refactoring of languages.  Still pickling issues to solve, but structure is complete.
