+++ revscoring/languages/english.py
@@ -112,6 +112,7 @@ informals = [
     r"b+o+n+e+r+",
     r"boobs?",
     r"bullshit",
+    r"bro",
     r"(bye)+",
     r"can'?t",
     r"[ck](oo+|e+w+)l+\w*",
@@ -130,9 +131,12 @@ informals = [
     r"m+e+o+w+",
     r"munch\w*",
     r"mom+(y|a)?",
+    r"moron",
     r"no+pe",
+    r"o+k+(a+y+)?",
     r"\w*o+m+g+\w*",
     r"poo+p\w*",
+    r"\w*retard\w*", r"tard",
     r"shove", r"smelly", r"soo+", r"stinky", r"\w*s+t+[uo]+p+i+d+\w*",
         r"suck(ing|er)?", r"sux", r"shouldn'?t",
     r"test +edit", r"t+u+r+d+s?\w*",
