+++ revscoring/languages/persian.py
