--- revscoring/new_language/__init__.py
@@ -1,1 +1,0 @@
-from .language import Language
