--- revscoring/new_language/language.py
@@ -1,5 +1,0 @@
-from types import ModuleType
-
-
-class Language(ModuleType):
-    pass
